Adding a new notification strategy "new failure and fixed".
It's similar to "failure and fixed" but does not notify of repeat failures.  This is useful because when a build fails and is set to rebuild automatically, you do not want to spam IRC with repeat failures.  The new strategy is also similar to "change", but change includes aborted builds and success after an aborted build, which we don't care about.  In particular, restarting jenkins while builds are taking place aborts the current builds, then rebuilds them on startup.  Even if the build succeeds, it will notify IRC that it went from Aborted to Success.