pom.xml: make Maven integration plugin optional
maven-plugin has been added as a dependency with 30ca89104bb along with
jenkins-war 2.3 and jenkins-core 2.3 and I guess it was no more bundled
in Jenkins as a result.

The only usage is in hudson.plugins.im.bot.HealthCommandTest, otherwhise
instant messaging does not depend on the plugin, it merely supports it.

Make the Maven plugin optional, this allows administrators to remove the
plugin when it is otherwise not used.