HealthCommandTest.java : restore tests
Two of the HealthCommandTest tests were ignored due to a mismatch in the
return value:

  HealthCommandTest.testFailure:87 WrongTypeOfReturnValue
  ItemGroup$MockitoMock$pmDFnSAx cannot be returned by getParent()
  getParent() should return Jenkins

The mock returns an ItemGroup when Mockito expects it to be Jenkins
instance. 70244952bc84d updated the parent pom which I guess updated
Mockito to a version that is stricter and check the returned value
types and the author went to ignore the tests.

The mocked getParent() method does not need to be mocked for the tests
to pass, thus remove the mocks and restore the tests.